third_party/WebKit/LayoutTests/fast/dom/computed-style-set-property-expected.txt

You should see two "PASS" messages below:

Computed style parent (should be null): null
PASS: got an exception when trying to change a computed style declaration, NoModificationAllowedError: Failed to set the 'color' property on 'CSSStyleDeclaration': These styles are computed, and therefore the 'color' property is read-only.

PASS: got an exception when trying to change a computed style declaration via setProperty, NoModificationAllowedError: Failed to execute 'setProperty' on 'CSSStyleDeclaration': These styles are computed, and therefore the 'color' property is read-only.
